personalized-real-estate-agent/requirements.txt
chromadb==0.5.0
langchain==0.2.2
openai==1.31.1
pydantic==2.7.3
tiktoken==0.7.0
tokenizers==0.19.1
transformers==4.41.2
gradio==4.33.0
langchain-openai==0.1.8
python-dotenv==1.0.1

requests~=2.32.3
pillow~=10.3.0
lancedb~=0.9.0
torch~=2.2.2